\docType{methods}
\name{covar}
\alias{covar}
\title{Identify Covariables}
\usage{
  covar(x)
}
\arguments{
  \item{x}{variable name}
}
\description{
  This is a special function used in the context of
  \code{\link{sanon}}. It identifies covariables when they
  appear on the right hand side of a formula.
}

