\docType{data}
\name{cpain}
\alias{cpain}
\title{Chronic Pain Data}
\format{A data frame with 193 observations and 4 variables}
\description{
  The data are from a multicenter randomized clinical trial
  to compare test and control treatments for the management
  of chronic pain, and they have had previous consideration
  in Stokes et al. (2000, chap. 13).
}
\details{
  \describe{ \item{\code{treat}}{a factor with levels
  \code{active} and \code{placebo} for treatment}
  \item{\code{response}}{a factor with five levels
  \code{poor}, \code{fair}, \code{moderate}, \code{good}
  and \code{excel} for pain status after treatment for 4
  weeks} \item{\code{center}}{a factor with two levels
  \code{I} and \code{II} for two centers}
  \item{\code{diagnosis}}{a factor with four levels
  \code{A}, \code{B}, \code{C}, and \code{D} for diagnoses}
  }
}
\references{
  Stokes, M. E., Davis, C. S., and Koch, G. G. (2000),
  Categorical Data Analysis using the SAS System, Cary: SAS
  Publishing.
}
\keyword{datasets}

