\docType{methods}
\name{grp}
\alias{grp}
\title{Identify Group Variables}
\usage{
  grp(x, ref = NULL)
}
\arguments{
  \item{x}{variable name}

  \item{ref}{character for the reference group for
  treatment group.}
}
\description{
  This is a special function used in the context of
  \code{\link{sanon}}. It identifies group variables when
  they appear on the right hand side of a formula.
}

