\docType{package}
\name{sanon-package}
\alias{sanon-package}
\title{The Stratified Analysis with Nonparametric covariable adjustment Package}
\description{
  A Package for Implementation of the method in Kawaguchi,
  Koch, and Wang (2011)
}
\author{
  Atsushi Kawaguchi. \email{kawa_a24@yahoo.co.jp}
}
\references{
  Kawaguchi, A., Koch, G. G., Wang, X. (2011): Stratified
  Multivariate Mann-Whitney Estimators for the Comparison
  of Two Treatments with Randomization Based Covariance
  Adjustment. Statistics in Biopharmaceutical Research,
  Vol. 3, No. 2, 217-231.
}
\seealso{
  \code{\link{sanon}}
}
\keyword{documentation}

