\docType{methods}
\name{strt}
\alias{strt}
\title{Identify Stratification Variables}
\usage{
  strt(x)
}
\arguments{
  \item{x}{variable name}
}
\description{
  This is a special function used in the context of
  \code{\link{sanon}}. It identifies stratification
  variables when they appear on the right hand side of a
  formula.
}

