% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sanon.r
\docType{methods}
\name{catecovar}
\alias{catecovar}
\title{Identify Categorical Covariables}
\usage{
catecovar(x, ref = NULL)
}
\arguments{
\item{x}{variable name}

\item{ref}{character for the reference group for the categorical covariable.}
}
\description{
This is a special function used in the context of \code{\link{sanon}}. It identifies categorical covariables when they appear on the right hand side of a formula.
}
\details{
In the \code{sanon}, the categorical covariable is converted into a dummy variable. The reference group is specified in the \code{ref} argument.
}

