\docType{methods}
\name{grp}
\alias{grp}
\title{Identify Group Variables}
\usage{
  grp(x, ref = NULL)
}
\arguments{
  \item{x}{variable name}

  \item{ref}{character for the reference group for
  treatment group.}
}
\description{
  This is a special function used in the context of
  \code{sanon}. It identifies group variables when they
  appear on the right hand side of a formula.
}
\seealso{
  Other sanon: \code{\link{catecovar}},
  \code{\link{confint.sanon}}, \code{\link{covar}},
  \code{\link{print.confint.sanon}},
  \code{\link{print.sanon}},
  \code{\link{print.summary.sanon}}, \code{\link{sanon}},
  \code{\link{sanon.default}}, \code{\link{sanon.formula}},
  \code{\link{strt}}, \code{\link{summary.sanon}}
}

