% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sanon.r
\docType{package}
\name{sanon-package}
\alias{sanon-package}
\title{The Stratified Analysis with Nonparametric covariable adjustment Package}
\description{
A Package for Implementation of the method in Kawaguchi, Koch, and Wang (2011)
}
\author{
Atsushi Kawaguchi. \email{kawa_a24@yahoo.co.jp}
}
\references{
Kawaguchi A., Koch, G. G. (2015). sanon: An R Package for Stratified Analysis with Nonparametric Covariable Adjustment. Journal of Statistical Software, 67(9), 1-37. doi:10.18637/jss.v067.i09

Kawaguchi A., Koch, G. G., Wang, X. (2011). Stratified Multivariate Mann-Whitney Estimators for the Comparison of Two Treatments with Randomization Based Covariance Adjustment. Statistics in Biopharmaceutical Research, Vol. 3, No. 2, 217-231.
}
\seealso{
\code{\link{sanon}}
}
\keyword{documentation}

