\docType{methods}
\name{strt}
\alias{strt}
\title{Identify Stratification Variables}
\usage{
  strt(x)
}
\arguments{
  \item{x}{variable name}
}
\description{
  This is a special function used in the context of
  \code{sanon}. It identifies stratification variables when
  they appear on the right hand side of a formula.
}
\seealso{
  Other sanon: \code{\link{catecovar}},
  \code{\link{confint.sanon}}, \code{\link{covar}},
  \code{\link{grp}}, \code{\link{print.confint.sanon}},
  \code{\link{print.sanon}},
  \code{\link{print.summary.sanon}}, \code{\link{sanon}},
  \code{\link{sanon.default}}, \code{\link{sanon.formula}},
  \code{\link{summary.sanon}}
}

